chore: Move files to separate Cloud Service Types for MySQLServers and PostgreSQLServers
