+++ src/plugin/manager/mysql_sql_flexible_servers/__init__.py
@@ -1,0 +1,1 @@
+from .server_manager import MySQLFlexibleServersManager
