+++ src/plugin/manager/postgre_sql_flexible_servers/server_manager.py
@@ -3,7 +3,7 @@ import logging
 from spaceone.inventory.plugin.collector.lib import *
 
 from plugin.conf.cloud_service_conf import ICON_URL
-from plugin.connector.postgre_sql_servers.postgresql_flexible_servers_connector import PostgreSQLFlexibleServersConnector
+from plugin.connector.postgre_sql_flexible_servers.postgresql_flexible_servers_connector import PostgreSQLFlexibleServersConnector
 from plugin.connector.subscriptions.subscriptions_connector import SubscriptionsConnector
 from plugin.manager.base import AzureBaseManager
 
@@ -11,8 +11,8 @@ _LOGGER = logging.getLogger(__name__)
 
 
 class PostgreSQLFlexibleServersManager(AzureBaseManager):
-    cloud_service_group = "PostgreSQLServers"
-    cloud_service_type = "FlexibleServer"
+    cloud_service_group = "PostgreSQLFlexibleServers"
+    cloud_service_type = "Server"
     service_code = "/Microsoft.DBforPostgreSQL/flexibleServers"
 
     def create_cloud_service_type(self):
