+++ src/plugin/manager/postgre_sql_servers/__init__.py
@@ -1,2 +1,1 @@
 from .server_manager import PostgreSQLServersManager
-from .flexible_server_manager import PostgreSQLFlexibleServersManager
