+++ src/plugin/metadata/my_sql_flexible_servers/server.yaml
