+++ src/plugin/metrics/MySQLFlexibleServers/Server/namespace.yaml
