+++ src/plugin/metrics/PostgreSQLFlexibleServers/Server/namespace.yaml
